backend/requirements.txt
# 后端 API 依赖（Render 部署）
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
httpx>=0.25.0
requests>=2.31.0
python-dotenv>=1.0.0
pillow>=10.0.0
volcengine>=1.0.0
numpy>=1.24.0

# 数据库相关
psycopg2-binary>=2.9.0  # PostgreSQL（Supabase 使用）
sqlalchemy>=2.0.0  # ORM
supabase>=2.0.0  # Supabase Python 客户端

# 对象存储
oss2>=2.18.0  # 阿里云 OSS
boto3>=1.28.0  # 亚马逊 S3

# 注意：RAG 相关依赖（opencv-python, sentence-transformers, chromadb）已移除
# 因为 RAG 服务不需要在 Render 后端运行，可以单独部署
